move callbacks to instrumentation
