+++ CHANGELOG.md
@@ -1,12 +1,15 @@
 # WaterDrop changelog
 
 ## 2.0.5 (Unreleased)
-- Fixes an issue where emited statistics from various producers would be publised to monitors of all the producers.
+- Fixes an issue where emited statistics from various producers would be published to monitors of all the producers.
 - Introduces support for error callbacks instrumentation notifications
 - Removes the `:producer` key from `statistics.emitted` and replaces it with `:producer_id`
 - Removes the `:producer` key from `message.acknowledged` and replaces it with `:producer_id`
 - Cleanup and refactor of callbacks support
 - Introduces a callbacks manager concept that will also be within in Karafka `2.0`
+- Sets default kafka `client.id` to `waterdrop`
+- Updates specs to always emit statistics for better test coverage
+- Adds statistics and errors integration specs running against Kafka
 
 ## 2.0.4 (2021-09-19)
 - Update `dry-*` to the recent versions and update settings syntax to match it
