+++ spec/lib/water_drop/instrumentation/callbacks/statistics_spec.rb
