+++ spec/lib/water_drop/instrumentation/callbacks_manager_spec.rb
