+++ spec/lib/water_drop/instrumentation/monitor_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe WaterDrop::Instrumentation::Monitor do
+RSpec.describe_current do
   subject(:monitor) { described_class.new }
 
   describe '#instrument' do
